Converted to Typescript
Converted all files to Typescript

Also simplified the API's to be a singular endpoint with aparameters.
