+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "azureFunctions.deploySubpath": "api",
     "azureFunctions.postDeployTask": "npm install",
-    "azureFunctions.projectLanguage": "JavaScript",
+    "azureFunctions.projectLanguage": "TypeScript",
     "azureFunctions.projectRuntime": "~2",
     "debug.internalConsoleOptions": "neverOpen",
     "azureFunctions.preDeployTask": "npm prune"
