+++ .vscode/tasks.json
@@ -6,7 +6,17 @@
 			"command": "host start",
 			"problemMatcher": "$func-watch",
 			"isBackground": true,
+			"dependsOn": "npm build",
+			"options": {
+				"cwd": "${workspaceFolder}/api"
+			}
+		},
+		{
+			"type": "shell",
+			"label": "npm build",
+			"command": "npm run build",
 			"dependsOn": "npm install",
+			"problemMatcher": "$tsc",
 			"options": {
 				"cwd": "${workspaceFolder}/api"
 			}
@@ -23,6 +33,7 @@
 			"type": "shell",
 			"label": "npm prune",
 			"command": "npm prune --production",
+			"dependsOn": "npm build",
 			"problemMatcher": [],
 			"options": {
 				"cwd": "${workspaceFolder}/api"
