+++ api/PersonManual/function.json
@@ -15,5 +15,6 @@
       "direction": "out",
       "name": "res"
     }
-  ]
+  ],
+  "scriptFile": "../dist/PersonManual/index.js"
 }
