+++ components/Icons.tsx
@@ -1,0 +1,253 @@
+import React from 'react';
+
+export function ReactIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 153 137"
+    >
+      <path
+        fill="#fff"
+        d="M152.549 68.351c0 11.446-11.213 22.13-28.891 28.153 3.53 17.712.632 31.99-9.046 37.552-9.947 5.719-24.767 1.761-38.834-10.495-13.711 11.761-27.569 16.389-37.238 10.81-9.939-5.734-12.961-20.98-9.333-39.246C10.794 89.058.494 79.947.494 68.351c0-11.18 11.234-20.851 28.674-26.835-3.906-19.021-.597-32.952 9.482-38.744 9.717-5.585 23.736-.718 37.649 11.35 14.474-12.809 28.354-16.719 38.423-10.91 9.708 5.6 12.478 20.138 8.936 38.178 17.082 5.936 28.891 15.827 28.891 26.961z"
+      ></path>
+      <path
+        fill="#2980B9"
+        d="M32.861 47.458a90.697 90.697 0 014.347-1.365 81.577 81.577 0 01-.671-2.955c-3.291-15.92-1.14-28.746 6.208-32.968 7.044-4.05 18.565.172 30.201 10.264a90.311 90.311 0 013.362 3.076 82.22 82.22 0 012.238-2.06c12.194-10.792 24.418-15.34 31.757-11.105 7.038 4.06 9.123 16.119 6.16 31.206a89.388 89.388 0 01-.998 4.458c1.73.49 3.4 1.012 4.998 1.567 14.297 4.968 23.427 12.754 23.427 20.83 0 8.341-9.801 16.708-24.693 21.78-1.207.41-2.425.79-3.651 1.138a88.566 88.566 0 011.061 4.805c2.824 14.826.618 26.598-6.401 30.634-7.25 4.166-19.418-.117-31.266-10.439a76.55 76.55 0 01-2.818-2.589 88.993 88.993 0 01-3.65 3.32c-11.477 9.843-22.812 13.818-29.825 9.771-7.243-4.179-9.597-16.825-6.54-32.211.232-1.176.504-2.376.81-3.598a79.366 79.366 0 01-2.512-.782C18.924 85.124 8.852 76.86 8.852 68.405c0-8.104 9.425-15.943 24.01-20.947z"
+      ></path>
+      <path
+        fill="#fff"
+        d="M36.22 84.778a75.32 75.32 0 002.267.707c1.71-5.392 4.016-11.127 6.838-17.05-2.693-5.784-4.91-11.445-6.572-16.803-1.38.4-2.722.82-4.014 1.263-12.496 4.287-20.119 10.627-20.119 15.51 0 5.204 8.233 11.959 21.599 16.373zm5.545 10.953c-1.351 6.804-1.544 12.955-.649 17.764.805 4.321 2.422 7.201 4.422 8.355 4.255 2.455 13.357-.736 23.17-9.154a82.68 82.68 0 003.397-3.085c-3.806-4.148-7.608-8.971-11.32-14.326-6.528-.577-12.696-1.521-18.289-2.81a70.97 70.97 0 00-.73 3.256zm56.088 25.695c4.158 1.464 7.469 1.505 9.471.355 4.26-2.448 6.03-11.901 3.615-24.58a82.184 82.184 0 00-.986-4.468c-5.531 1.22-11.654 2.097-18.198 2.626-3.736 5.24-7.649 10.056-11.592 14.274a71.012 71.012 0 002.573 2.361c5.24 4.565 10.49 7.803 15.116 9.432h.001zm19.479-36.68c6.584-2.243 12.022-5.158 15.75-8.34 3.349-2.858 5.04-5.697 5.04-8 0-4.9-7.331-11.152-19.558-15.4a83.596 83.596 0 00-4.648-1.458c-1.69 5.478-3.907 11.206-6.582 16.998 2.71 5.878 4.958 11.697 6.664 17.24 1.12-.32 2.231-.666 3.334-1.04zm-6.53-44.299c2.538-12.927.853-22.679-3.389-25.126-4.518-2.606-14.507 1.11-25.036 10.428a75.566 75.566 0 00-2.026 1.865c3.923 4.199 7.8 8.979 11.503 14.188 6.351.587 12.431 1.53 18.02 2.791.344-1.374.653-2.757.928-4.146zm-58.25 14.336c1.32-2.273 2.69-4.517 4.107-6.731a127.193 127.193 0 00-12.313 2.145c1.164 3.717 2.614 7.604 4.322 11.59a200.317 200.317 0 013.885-7.004zm23.755-23.06c-2.661 2.873-5.326 6.08-7.947 9.561 5.317-.25 10.643-.252 15.96-.003a126.902 126.902 0 00-8.013-9.558zm23.906 23.099c1.325 2.29 2.596 4.61 3.812 6.96a126.674 126.674 0 004.291-11.646c-3.853-.86-7.962-1.562-12.243-2.095a169.004 169.004 0 014.14 6.781zM95.95 89.181c4.423-.492 8.593-1.159 12.445-1.994-1.192-3.837-2.661-7.826-4.377-11.882A169.288 169.288 0 0195.95 89.18zm-19.803 16.314c2.734-2.94 5.461-6.193 8.125-9.69a201.91 201.91 0 01-7.9.152c-2.754 0-5.475-.062-8.153-.18 2.63 3.561 5.285 6.818 7.928 9.718zM48.613 75.098c-1.805 4.1-3.327 8.068-4.535 11.84 3.916.89 8.143 1.608 12.597 2.14a200.82 200.82 0 01-4.146-6.849 182.492 182.492 0 01-3.916-7.131zm8.914 4.259a188.454 188.454 0 006.478 10.418c4.007.285 8.147.432 12.366.432 4.202 0 8.29-.13 12.227-.384a163.04 163.04 0 0012.36-21.279 162.79 162.79 0 00-5.744-10.85 161.522 161.522 0 00-6.547-10.389 163.98 163.98 0 00-24.615.004 173.11 173.11 0 00-6.507 10.354 188.334 188.334 0 00-5.821 10.8 173.042 173.042 0 005.802 10.894zM45.626 15.15c-4.522 2.6-6.28 13.082-3.44 26.827.187.896.39 1.788.608 2.676 5.602-1.288 11.686-2.247 18.055-2.843 3.711-5.266 7.556-10.055 11.415-14.2a84.742 84.742 0 00-3.105-2.838c-9.97-8.648-19.287-12.062-23.533-9.622z"
+      ></path>
+      <path
+        fill="#2980B9"
+        d="M76.371 56.387c-6.66 0-12.058 5.382-12.058 12.019 0 6.638 5.398 12.019 12.058 12.019s12.059-5.381 12.059-12.02c0-6.636-5.4-12.018-12.059-12.018"
+      ></path>
+    </svg>
+  );
+}
+
+export function VueIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 153 132"
+    >
+      <g clipPath="url(#clip0)">
+        <path
+          fill="#3498DB"
+          d="M122.388 0h30.598L76.493 131.631 0 0h58.517l17.976 30.523L94.086 0h28.302z"
+        ></path>
+        <path
+          fill="#3498DB"
+          d="M0 0l76.493 131.631L152.986 0h-30.598L76.493 78.978 30.215 0H0z"
+        ></path>
+        <path
+          fill="#2980B9"
+          d="M30.215 0l46.278 79.36L122.388 0H94.086L76.493 30.523 58.517 0H30.215z"
+        ></path>
+      </g>
+      <defs>
+        <clipPath id="clip0">
+          <path fill="#fff" d="M0 0H153V131.75H0z"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export function SvelteIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      viewBox="0 0 107 128"
+    >
+      <path
+        fill="#2980B9"
+        d="M94.157 22.819c-10.4-14.885-30.94-19.297-45.792-9.835L22.282 29.608A29.923 29.923 0 008.764 49.65a31.514 31.514 0 003.108 20.231 30.006 30.006 0 00-4.477 11.183 31.889 31.889 0 005.448 24.116c10.402 14.886 30.942 19.297 45.791 9.835l26.083-16.624A29.918 29.918 0 0098.235 78.35a31.526 31.526 0 00-3.105-20.232 30 30 0 004.474-11.182 31.88 31.88 0 00-5.447-24.116"
+      ></path>
+      <path
+        fill="#fff"
+        d="M45.817 106.582a20.718 20.718 0 01-22.237-8.243 19.174 19.174 0 01-3.277-14.503 18.189 18.189 0 01.624-2.435l.49-1.498 1.337.981a33.644 33.644 0 0010.203 5.098l.97.294-.09.968a5.847 5.847 0 001.052 3.878 6.239 6.239 0 006.695 2.485 5.745 5.745 0 001.603-.704L69.27 76.28a5.43 5.43 0 002.45-3.631 5.795 5.795 0 00-.987-4.371 6.244 6.244 0 00-6.698-2.487 5.743 5.743 0 00-1.6.704l-9.953 6.345a19.033 19.033 0 01-5.296 2.326 20.718 20.718 0 01-22.237-8.243 19.172 19.172 0 01-3.277-14.502 17.988 17.988 0 018.13-12.052l26.081-16.623a19.004 19.004 0 015.3-2.329 20.718 20.718 0 0122.237 8.243 19.174 19.174 0 013.277 14.503 18.4 18.4 0 01-.624 2.435l-.49 1.498-1.337-.98a33.617 33.617 0 00-10.203-5.1l-.97-.294.09-.968a5.859 5.859 0 00-1.052-3.878 6.239 6.239 0 00-6.696-2.485 5.745 5.745 0 00-1.602.704L37.73 51.72a5.422 5.422 0 00-2.449 3.63 5.786 5.786 0 00.986 4.372 6.244 6.244 0 006.698 2.486 5.765 5.765 0 001.602-.704l9.952-6.342a18.978 18.978 0 015.295-2.328 20.718 20.718 0 0122.237 8.242 19.172 19.172 0 013.277 14.503 17.998 17.998 0 01-8.13 12.053l-26.081 16.622a19.004 19.004 0 01-5.3 2.328"
+      ></path>
+    </svg>
+  );
+}
+
+export function NextIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 153 93"
+    >
+      <g clipPath="url(#clip0)">
+        <path
+          fill="#2980B9"
+          d="M36.168 24.086h28.81v2.295H38.81v17.271h24.607v2.295H38.81V64.91h26.468v2.295h-29.11V24.086zm31.39 0h3.062l13.564 18.962 13.864-18.962L116.906 0 85.925 45.041l15.965 22.164h-3.182l-14.524-20.17L69.6 67.204h-3.122L82.563 45.04 67.56 24.086zm35.472 2.295v-2.295h32.83v2.295h-15.124v40.823h-2.641V26.38H103.03zM.127 24.086h3.301l45.52 68.265-18.812-25.146L2.888 27.347l-.12 39.858H.128V24.086zm135.468 40.13c-.541 0-.945-.418-.945-.96 0-.541.404-.96.944-.96.546 0 .945.419.945.96 0 .542-.399.96-.945.96h.001zm2.594-2.526h1.414c.02.767.579 1.282 1.401 1.282.918 0 1.438-.553 1.438-1.591v-6.574h1.439v6.58c0 1.87-1.079 2.947-2.865 2.947-1.676 0-2.826-1.045-2.826-2.644h-.001zm7.573-.083h1.427c.122.882.982 1.443 2.222 1.443 1.155 0 2.004-.6 2.004-1.424 0-.709-.539-1.135-1.767-1.424l-1.195-.29c-1.675-.393-2.44-1.206-2.44-2.572 0-1.657 1.348-2.76 3.371-2.76 1.883 0 3.258 1.103 3.341 2.669h-1.4c-.135-.857-.88-1.392-1.959-1.392-1.138 0-1.895.548-1.895 1.386 0 .664.488 1.044 1.696 1.328l1.021.25c1.901.446 2.685 1.22 2.685 2.618 0 1.779-1.376 2.895-3.571 2.895-2.056 0-3.437-1.065-3.54-2.728v.001z"
+        ></path>
+      </g>
+      <defs>
+        <clipPath id="clip0">
+          <path fill="#fff" d="M0 0H153V92.351H0z"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export function NuxtIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 153 153"
+    >
+      <path
+        fill="#3498DB"
+        d="M87.18 51.045l-5.236 9.219-17.907-31.536-54.97 96.797h37.218c0 5.092 4.06 9.219 9.067 9.219H9.067c-3.24 0-6.232-1.758-7.851-4.61a9.353 9.353 0 010-9.218l54.97-96.797c1.62-2.853 4.614-4.61 7.853-4.61 3.24 0 6.233 1.757 7.852 4.61L87.18 51.044z"
+      ></path>
+      <path
+        fill="#2F495E"
+        d="M126.862 120.916L92.416 60.264l-5.236-9.219-5.236 9.219-34.442 60.652a9.353 9.353 0 000 9.218c1.618 2.852 4.611 4.61 7.85 4.61H119c3.24 0 6.233-1.756 7.854-4.608a9.353 9.353 0 000-9.22h.008zm-71.51 4.609L87.18 69.483 119 125.525H55.352z"
+      ></path>
+      <path
+        fill="#2980B9"
+        d="M151.489 130.135c-1.619 2.851-4.612 4.609-7.85 4.609H119c5.007 0 9.067-4.127 9.067-9.219h15.553L99.488 47.807l-7.072 12.457-5.236-9.219 4.458-7.847c1.62-2.853 4.612-4.61 7.852-4.61 3.239 0 6.232 1.757 7.852 4.61l44.147 77.718a9.356 9.356 0 010 9.219z"
+      ></path>
+    </svg>
+  );
+}
+
+export function GatsbyIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 153 153"
+    >
+      <path
+        fill="#2980B9"
+        d="M76.5 0C34.248 0 0 34.248 0 76.5S34.248 153 76.5 153 153 118.752 153 76.5 118.752 0 76.5 0zM16.438 77.296l59.266 59.266c-32.546-.423-58.843-26.72-59.266-59.266zm73.508 57.766L17.938 63.054C24.046 36.353 47.944 16.426 76.5 16.426c19.957 0 37.644 9.737 48.568 24.719l-8.317 7.338C107.889 35.778 93.165 27.46 76.5 27.46c-21.218 0-39.29 13.477-46.12 32.338l62.822 62.816c15.257-5.526 26.989-18.403 30.918-34.351H98.085V76.5h38.495c-.006 28.556-19.933 52.455-46.634 58.562z"
+      ></path>
+    </svg>
+  );
+}
+
+export function StarIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 16 16"
+    >
+      <g clip-path="url(#clip0)">
+        <path
+          fill="#3498DB"
+          d="M3.581 15.677a.85.85 0 01-.829-1.04l1.03-4.535L.29 7.041a.85.85 0 01.482-1.488l4.619-.42L7.218.86a.85.85 0 011.564 0l1.826 4.273 4.618.42a.85.85 0 01.483 1.488l-3.49 3.06 1.029 4.535a.85.85 0 01-1.265.918L8 13.174l-3.983 2.382a.852.852 0 01-.436.12zM8 12.149c.151 0 .301.04.436.12l3.759 2.249-.972-4.28a.85.85 0 01.27-.828l3.296-2.891-4.361-.396a.85.85 0 01-.707-.516L8 1.575 6.277 5.608a.847.847 0 01-.705.514l-4.362.396 3.296 2.89c.237.207.34.524.27.83l-.97 4.279 3.758-2.247A.853.853 0 018 12.149zM5.357 5.216l-.001.001zm5.285-.002v.001z"
+        ></path>
+      </g>
+      <defs>
+        <clipPath id="clip0">
+          <path fill="#fff" d="M0 0H16V16H0z"></path>
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+
+export function WatchIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 18 18"
+    >
+      <path
+        fill="#3498DB"
+        d="M17.886 8.65C17.725 8.43 13.893 3.264 9 3.264 4.106 3.264.275 8.43.114 8.65a.594.594 0 000 .7c.16.22 3.992 5.386 8.886 5.386 4.893 0 8.725-5.166 8.886-5.386a.593.593 0 000-.7zM9 13.55c-3.605 0-6.727-3.43-7.65-4.55C2.271 7.878 5.387 4.45 9 4.45c3.604 0 6.726 3.429 7.65 4.55-.922 1.122-4.038 4.55-7.65 4.55z"
+      ></path>
+      <path
+        fill="#3498DB"
+        d="M9 5.44A3.565 3.565 0 005.44 9 3.565 3.565 0 009 12.56 3.565 3.565 0 0012.56 9 3.565 3.565 0 009 5.44zm0 5.934A2.376 2.376 0 016.626 9 2.376 2.376 0 019 6.626 2.376 2.376 0 0111.373 9 2.376 2.376 0 019 11.374z"
+      ></path>
+    </svg>
+  );
+}
+
+export function BugIcon({ w, h }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width={w}
+      height={h}
+      fill="none"
+      viewBox="0 0 18 18"
+    >
+      <path
+        fill="#E74C3C"
+        d="M17.48 9.173h-3.115v-1.08a3.294 3.294 0 002.77-3.247v-.692a.52.52 0 00-1.039 0v.692c0 1.062-.74 1.954-1.73 2.19v-.459a.866.866 0 00-.866-.865h-.52V3.808A2.254 2.254 0 0011.166 1.6l.386-.87a.52.52 0 00-.949-.421l-.555 1.249H7.953L7.398.308a.52.52 0 00-.95.422l.387.87A2.254 2.254 0 005.02 3.808v1.904H4.5a.866.866 0 00-.865.865v.458a2.253 2.253 0 01-1.731-2.189v-.692a.52.52 0 00-1.039 0v.692a3.294 3.294 0 002.77 3.247v1.08H.519a.52.52 0 000 1.038h3.116v1.08a3.294 3.294 0 00-2.77 3.247v.693a.52.52 0 001.039 0v-.693c0-1.062.74-1.953 1.73-2.188v.284A5.372 5.372 0 009 18a5.371 5.371 0 005.365-5.365v-.285a2.254 2.254 0 011.731 2.189v.692a.52.52 0 001.038 0v-.692a3.294 3.294 0 00-2.769-3.248v-1.08h3.116a.52.52 0 000-1.038zM6.059 3.808c0-.668.543-1.212 1.211-1.212h3.462c.668 0 1.211.544 1.211 1.212v1.904H6.058V3.808zm7.269 8.827A4.332 4.332 0 019 16.962a4.332 4.332 0 01-4.327-4.327V6.75h3.808v5.02a.52.52 0 001.038 0V6.75h3.808v5.885z"
+      ></path>
+    </svg>
+  );
+}
+
+export function AzureIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="20"
+      fill="none"
+      viewBox="0 0 24 20"
+    >
+      <path
+        fill="#fff"
+        d="M13.113 0L6.037 6.15 0 17.016h5.444L13.113 0zm.94 1.44L11.035 9.97l5.79 7.292-11.232 1.935H24L14.054 1.44z"
+      ></path>
+    </svg>
+  );
+}
+
+export function GithubIcon() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      fill="none"
+      viewBox="0 0 24 24"
+    >
+      <path
+        fill="#fff"
+        d="M24 12c0-3.313-1.172-6.14-3.516-8.484C18.141 1.172 15.312 0 12 0 8.688 0 5.86 1.172 3.516 3.516 1.172 5.859 0 8.687 0 12c0 3.313 1.172 6.14 3.516 8.484C5.859 22.828 8.687 24 12 24c3.313 0 6.14-1.172 8.484-3.516C22.828 18.141 24 15.312 24 12zm-1.313 0c0 2.938-1.046 5.453-3.14 7.547-2.094 2.094-4.61 3.14-7.547 3.14-2.938 0-5.453-1.046-7.547-3.14-2.094-2.094-3.14-4.61-3.14-7.547 0-2.938 1.046-5.453 3.14-7.547C6.547 2.36 9.063 1.313 12 1.313c2.938 0 5.453 1.046 7.547 3.14 2.094 2.094 3.14 4.61 3.14 7.547zm-3.375 2.578c.063-.062.063-.125 0-.187a.507.507 0 00-.468-.141c-.188.031-.297.063-.328.094-.22.094-.461.304-.727.633-.266.328-.477.554-.633.68-.75.624-1.515.671-2.297.14a1.404 1.404 0 00-.304-.563 1.42 1.42 0 00-.399-.328 9.925 9.925 0 00-.656-.281c1.031-.094 1.867-.274 2.508-.54.64-.265 1.148-.6 1.523-1.007.5-.562.805-1.297.914-2.203.11-.906.04-1.734-.21-2.484a4.128 4.128 0 00-.797-1.266c.25-.781.203-1.672-.141-2.672-1 .063-1.906.406-2.719 1.031-1.562-.406-3.187-.422-4.875-.047a8.81 8.81 0 00-1.008-.585c-.453-.235-1.023-.368-1.71-.399-.126.344-.211.75-.258 1.219-.047.469-.024.906.07 1.312-.719.75-1.094 1.75-1.125 3 0 1 .172 1.828.516 2.485.343.656.953 1.203 1.828 1.64.593.282 1.375.453 2.343.516-.437.219-.757.445-.96.68-.204.234-.336.601-.399 1.101v1.477l-.047 1.43c-.094.156-.203.289-.328.398-.125.11-.234.187-.328.234-.094.047-.125.117-.094.211.031.094.156.14.375.14.531 0 .953-.187 1.266-.562.094-.156.14-.343.14-.562v-2.25c0-.25.063-.43.188-.54.125-.109.25-.179.375-.21v2.953c0 .531-.063.875-.188 1.031a1.104 1.104 0 00-.234.469c-.031.031-.008.055.07.07.078.016.211-.008.399-.07.406-.094.687-.305.844-.633.156-.328.234-.68.234-1.055v-2.906h.61v2.907c0 .375.077.734.234 1.078.218.437.625.672 1.218.703.094-.032.141-.063.141-.094-.031-.031-.125-.188-.281-.469a2.723 2.723 0 01-.188-1.031v-2.953a.759.759 0 01.352.21c.11.11.164.29.164.54v2.25c0 .219.047.406.14.562.282.375.704.563 1.266.563.219 0 .344-.047.375-.14.031-.063.024-.118-.023-.165a1.424 1.424 0 00-.211-.164 1.408 1.408 0 01-.188-.14.852.852 0 01-.281-.375c-.094-.125-.117-.454-.07-.985.046-.531.054-.875.023-1.031.813.281 1.64.14 2.485-.422.25-.25.5-.61.75-1.078.187-.344.562-.75 1.125-1.219z"
+      ></path>
+    </svg>
+  );
+}
+
+export const projectIcons = {
+  react: ReactIcon,
+  vue: VueIcon,
+  svelte: SvelteIcon,
+  next: NextIcon,
+  nuxt: NuxtIcon,
+  gatsby: GatsbyIcon
+}
