+++ package.json
@@ -14,6 +14,9 @@
     "react-dom": "16.13.1"
   },
   "devDependencies": {
-    "serve": "^11.3.0"
+    "@types/node": "^14.0.23",
+    "@types/react": "^16.9.43",
+    "serve": "^11.3.0",
+    "typescript": "^3.9.6"
   }
 }
