--- pages/_app.js
@@ -1,11 +1,0 @@
-import '../styles/reset.css';
-import '../styles/main.css';
-
-// This default export is required in a new `pages/_app.js` file.
-export default function MyApp({ Component, pageProps }) {
-  return (
-    <div>
-      <Component {...pageProps} />
-    </div>
-  );
-}
