--- utils/projectsData.js
@@ -1,40 +1,0 @@
-const projects = [
-  {
-    id: "react",
-    name: "React",
-    path: "facebook/react",
-    slug: "facebook-react",
-  },
-  {
-    id: "vue",
-    name: "Vue",
-    path: "vuejs/vue",
-    slug: "vuejs-vue",
-  },
-  {
-    id: "svelte",
-    name: "Svelte",
-    path: "sveltejs/svelte",
-    slug: "sveltejs-svelte",
-  },
-  {
-    id: "next",
-    name: "Next.js",
-    path: "zeit/next.js",
-    slug: "zeit-nextjs",
-  },
-  {
-    id: "nuxt",
-    name: "Nuxt.js",
-    path: "nuxt/nuxt.js",
-    slug: "nuxt-nuxtjs",
-  },
-  {
-    id: "gatsby",
-    name: "Gatsby",
-    path: "gatsbyjs/gatsby",
-    slug: "gatsbyjs-gatsby",
-  },
-];
-
-module.exports = { projects };
